refactor filtering of events into FindCommand
